+++ packages/json/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.10.0 — not yet released
 
 * Add a `LionWebJsonDeltaChunk` type for the delta protocol.
-    (Also add a `NodesOfLionWebJsonChunk` type to refactor to DRY.)
+    (Also add a `OnlyNodesOfLionWebJsonChunk` type to refactor to DRY.)
 
 
 ## 0.9.0